requirements.txt
django==2.2.5
djangorestframework==3.12.4
djoser==2.1.0
django-environ==0.8.1
Pillow==8.4.0